RFC: basic extraction to a library.
no work done to refactor the existing -e -t modes to use the library more, just
proof of concept at this stage.

not sure which parts of the parser api should really be exposed, and how much
it should be mangled to support the existing export style commands
